HT-2467 normalize enumchrons as module for holdings   * identical tests as ht_items   * moves the duplicated enumchron handling into its own module
